% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unc_core.R
\name{aggregate_by_group}
\alias{aggregate_by_group}
\title{Aggregate emissions uncertainties by groups}
\usage{
aggregate_by_group(dt_emissions, by_group, correlated = TRUE)
}
\arguments{
\item{dt_emissions}{Data table of emissions}
}
\value{
A data table
}
\description{
Aggregates a table of emissions uncertainties based on a specified group.
}
