% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unc_core.R
\name{aggregate_substance}
\alias{aggregate_substance}
\title{Aggregate emissions uncertainties by substance}
\usage{
aggregate_substance(
  dt_emissions,
  correlate_in_fuel = TRUE,
  correlate_in_sector = TRUE
)
}
\arguments{
\item{dt_emissions}{Data table of emissions}
}
\value{
A data table
}
\description{
Aggregates a table of emissions uncertainties depending on whether CO2 or
not. If CO2 aggregates by fuel. Otherwise, aggregates by sector. In both cases
the emissions within these groups are considered as correlated by default.
}
