% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unc_core.R
\name{aggregate_unc_dt}
\alias{aggregate_unc_dt}
\title{Aggregate a table of emissions uncertainties}
\usage{
aggregate_unc_dt(dt, col_emissions, col_min, col_max, col_group)
}
\arguments{
\item{dt}{A data table of emissions}

\item{col_emissions}{Column name where the emissions data is in}

\item{col_min}{Column name of max uncertainty}

\item{col_max}{Column name of min uncertainty}

\item{col_group}{Column to group by}
}
\value{
A data table of aggregated uncertainties by group
}
\description{
One-shot aggregation of emissions uncertainties using data.table.
}
