% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_sectors}
\alias{check_sectors}
\title{Check whether sector codes are valid}
\usage{
check_sectors(sectors)
}
\arguments{
\item{sectors}{Character vector of sector codes}
}
\value{
Nothing, or an error message
}
\description{
Check against stored top-level EDGAR sector codes. This is to avoid sending incorrect queries
to the DB which might be harder to debug.
}
\details{
Note that these codes are taken from EDGAR and stored
in the package to avoid having to run the query each time. Occasionally, you
might need to update this: do so by running `get_valid_sectors()` to retrieve the vector of ISO codes, then:

`usethis::use_data(valid_sectors, overwrite = TRUE)`

to store in the package. This will only work if you are working in the package
source directory.
}
