% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_substances}
\alias{check_substances}
\title{Check whether substance codes are valid}
\usage{
check_substances(substances)
}
\arguments{
\item{substances}{Character vector of uncertainty codes}
}
\value{
Nothing, or an error message
}
\description{
Just checks whether they are one of the substances where we have uncertainty
data available.
}
