% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_years}
\alias{check_years}
\title{Check whether years are valid}
\usage{
check_years(years)
}
\arguments{
\item{years}{Numerical vector of years}
}
\value{
Nothing, or an error message
}
\description{
Are the user-input years in the valid range?
}
