% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{connect_to_edgar}
\alias{connect_to_edgar}
\title{Establish connection with EDGAR DB}
\usage{
connect_to_edgar()
}
\value{
ODBC connection
}
\description{
You need to be on the network for this to work. The user ID and pwd should
be stored in an R file which loads these strings as variables `uname` and `pwd`.
}
