% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unc_core.R
\name{f_agg_correlated}
\alias{f_agg_correlated}
\title{Aggregate correlated emissions}
\usage{
f_agg_correlated(emi, u)
}
\arguments{
\item{emi}{Vector of emissions values}

\item{u}{Vector of emissions uncertainties as percentages or fractions}
}
\value{
Percentage or fraction uncertainty of total emissions
}
\description{
Aggregates a vector of uncertainties and emissions using correlated formula.
}
