% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unc_core.R
\name{f_agg_uncorrelated}
\alias{f_agg_uncorrelated}
\title{Aggregate uncorrelated emissions}
\usage{
f_agg_uncorrelated(emi, u)
}
\arguments{
\item{emi}{Vector of emissions values}

\item{u}{Vector of emissions uncertainties as percentages or fractions}
}
\value{
Percentage or fraction uncertainty of total emissions
}
\description{
Aggregates a vector of uncertainties and emissions using uncorrelated formula.
}
