% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_country_info}
\alias{get_country_info}
\title{Get country info table}
\usage{
get_country_info(con)
}
\arguments{
\item{con}{ODBC connection to EDGAR: use \code{\link[=connect_to_edgar]{connect_to_edgar()}}}
}
\value{
data.table of country info
}
\description{
Retrieves the "Countries" table from EDGAR, which includes the development
group of each country (needed for uncertainty calculation).
}
