% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_level.R
\name{get_uncertain_emissions}
\alias{get_uncertain_emissions}
\title{Get aggregated uncertain emissions}
\usage{
get_uncertain_emissions(
  con,
  substances = NULL,
  years = NULL,
  countries = NULL,
  sectors = NULL,
  agg_countries = FALSE,
  agg_sectors = TRUE,
  agg_substance = FALSE
)
}
\arguments{
\item{con}{DB connection from [connect_to_edgar()]}

\item{substances}{One or more of "CO2", "N2O", "CH4"}

\item{years}{Vector of years}

\item{countries}{Vector of ISO3 country codes}

\item{sectors}{Vector of EDGAR sectors}

\item{agg_countries}{Logical: whether to aggregate all countries together}

\item{agg_sectors}{Logical: whether to aggregate all sectors together}

\item{agg_substance}{Logical: whether to aggregate substances together - NOTE
that this is not properly implemented since the emissions are in different
units at the moment.}
}
\value{
Data frame of emissions
}
\description{
Returns a table of emissions data with confidence intervals, aggregated up to
a specified limit.
}
