% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_uncertainty_table}
\alias{get_uncertainty_table}
\title{Get uncertainty table from DB}
\usage{
get_uncertainty_table(con)
}
\arguments{
\item{con}{DB connection established with [connect_to_edgar()].}
}
\value{
Data frame
}
\description{
Returns the uncertainty table giving uncertainty values for emissions at the
lowest level.
}
