% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iplots.R
\name{iplot_time_series}
\alias{iplot_time_series}
\title{Plot time series of emissions}
\usage{
iplot_time_series(dt_emissions)
}
\arguments{
\item{dt_emissions}{}
}
\value{
Plotly plot
}
\description{
Plots a time series of emissions, with confidence intervals. It takes a
data frame with expected columns "Year", "Emissions", "Emissions_Min",
"Emissions_Max".
}
