% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proc_code_depth}
\alias{proc_code_depth}
\title{Get depth of process code}
\usage{
proc_code_depth(codes)
}
\arguments{
\item{codes}{Vector of codes to test.}
}
\value{
Value between 1 and 5
}
\description{
Checks the format of a vector of EDGAR process codes, ensures they are valid,
and returns the depth (value between 1 and 5), or an error if codes are not
in the expected format.
}
