% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shrink_process_codes}
\alias{shrink_process_codes}
\title{Shrink process codes}
\usage{
shrink_process_codes(codes, depth = 3)
}
\arguments{
\item{codes}{A vector of process codes, like "ENE.AEL.CRU.GT0.000"}

\item{depth}{Depth to shrink to (value between 1 and 5)}
}
\value{
Vector of length `codes` but shrunk
}
\description{
Shortens EDGAR process codes to a specified depth
}
