% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unc_core.R
\name{aggregate_unc}
\alias{aggregate_unc}
\title{Aggregate a set of emissions uncertainties}
\usage{
aggregate_unc(dt_emissions, correlated = FALSE)
}
\arguments{
\item{dt_emissions}{Data table with columns Emissions, Emissions_Min, Emissions_Max}

\item{correlated}{If `TRUE` aggregates emissions assuming full correlation.}
}
\value{
A scalar value
}
\description{
Note that `NA`s in emissions or emissions uncertainties are excluded.
}
