% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unc_core.R
\name{aggregate_unc}
\alias{aggregate_unc}
\title{Aggregate a set of emissions uncertainties}
\usage{
aggregate_unc(emissions_uncertainties, emissions)
}
\arguments{
\item{emissions_uncertainties}{A vector of emissions uncertainties, as e.g.
fractions or percentages. Must be same length as `emissions`.}

\item{emissions}{A vector of emissions values}
}
\value{
A scalar value
}
\description{
Note that `NA`s in emissions or emissions uncertainties are excluded.
}
