% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unc_core.R
\name{f_agg_correlated}
\alias{f_agg_correlated}
\title{Aggregate a table of emissions uncertainties}
\usage{
f_agg_correlated(emi, u)
}
\arguments{
\item{dt}{A data table of emissions}

\item{col_emissions}{Column name where the emissions data is in}

\item{col_min}{Column name of max uncertainty}

\item{col_max}{Column name of min uncertainty}

\item{col_group}{Column to group by}
}
\value{
A data table of aggregated uncertainties by group
}
\description{
One-shot aggregation of emissions uncertainties using data.table.
}
