% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_emissions_data}
\alias{get_emissions_data}
\title{Get emissions data}
\usage{
get_emissions_data(
  con,
  substances = NULL,
  countries = NULL,
  years = NULL,
  sectors = NULL,
  emi_id = "29072022103026",
  table_name = "emi_edgar_release"
)
}
\arguments{
\item{con}{An ODBC connection: use \code{\link[=connect_to_edgar]{connect_to_edgar()}}}

\item{substances}{A character vector containing any of "CO2", "N2O", "CH4". Or if \code{NULL}
returns all substances.}

\item{countries}{A character vector of countries to return data from. Or if \code{NULL} returns
all countries.}

\item{years}{Vector of years within the range 1960:2100}

\item{sectors}{Character vector of sector codes to return}

\item{emi_id}{data set ID, as a string}

\item{table_name}{Name of the table to query - if not the default the query might not work}
}
\value{
A data.table of emissions data
}
\description{
Returns a data.table with emissions data as specified. Obviously depending on the query this
can lead to a very large table, so use carefully. The emissions data is returned as the lowest
level available in EDGAR, without aggregating to higher levels.
}
